feat: add support for Namespace Objects
This adds the ability for the commands that take a namespace to take a namespace object. This allows for `Get-KubeNamespace | Get-KubePod` to be used and other situations.

While we are here, clean up the naming convention and leave an alias behind to prevent backwards compatibility problems. Additionally, use const strings for ParameterSetNames to give us a compiler warning if there is a typo.

These changes try to follow how Azure PowerShell is structured.